fix(index): Don't talk to crates.io when using alt registry
Ideally we'd find the proper registry and talk to it but that takes a
lot more work to support.
Until then, we'll just act like the registry has no packages in it and
put in some specific hacks to support it.

This should be less brittle than our previous approach which just hoped
we sprinkled `config.registr().is_none()` in the right places.


Fixes #732